perf: use aho-corasick to check lines
