+++ src/rust/src/kractor/output.rs
@@ -1,53 +1,16 @@
-use std::collections::HashSet;
 use std::fmt::Display;
 use std::fs::File;
-use std::io::{BufRead, BufReader, BufWriter, Read, Write};
+use std::io::{BufWriter, Read, Write};
 use std::path::Path;
 
+use aho_corasick::AhoCorasick;
 use crossbeam_channel::bounded;
 use extendr_api::prelude::*;
-use memchr::memmem;
-use noodles_fasta::io::Writer;
-use noodles_fasta::record::{definition, sequence, Record};
-use noodles_fastq::io::Reader;
-
-#[extendr]
-#[allow(clippy::too_many_arguments)]
-fn kractor(
-    koutput: &str,
-    patterns: Robj,
-    ofile: &str,
-    fq1: &str,
-    ofile1: &str,
-    fq2: Option<&str>,
-    ofile2: Option<&str>,
-    io_buffer: usize,
-    buffersize: usize,
-    batchsize: usize,
-    queue_capacity: usize,
-    threads: usize,
-) -> std::result::Result<(), String> {
-    let patterns: Vec<&str> = patterns
-        .as_str_vector()
-        .ok_or("`patterns` must be a character vector")?;
-    write_matching_output(
-        koutput,
-        &patterns,
-        ofile,
-        io_buffer,
-        buffersize,
-        batchsize,
-        queue_capacity,
-        threads,
-    )?;
-    let id_set = read_sequence_id_from_koutput(ofile, io_buffer)?;
-    write_matching_reads(fq1, ofile1, fq2, ofile2, &id_set, io_buffer)
-}
 
 #[allow(clippy::too_many_arguments)]
-fn write_matching_output<P>(
+pub fn write_matching_output<P>(
     koutput: P,
-    patterns: &[&str],
+    needles: &AhoCorasick,
     ofile: P,
     io_buffer: usize,
     buffersize: usize,
@@ -66,11 +29,6 @@ where
         io_buffer,
         File::create(ofile).map_err(|e| e.to_string())?,
     );
-    let patterns: Vec<_> = patterns
-        .iter()
-        .map(|s| memmem::Finder::new(s.as_bytes()))
-        .collect();
-    let needles = &patterns;
 
     // Start the processor threads
     let (writer_tx, ref writer_rx) =
@@ -106,14 +64,10 @@ where
                         while let Some(pos) =
                             chunk.iter().position(|&b| b == b'\n')
                         {
-                            // SAFETY: pos is always in chunk
                             let line: Vec<u8> = chunk.drain(..= pos).collect();
                             let mut fields = line.splitn(3, |b| *b == b'\t');
                             if let Some(taxid) = fields.nth(2) {
-                                if needles
-                                    .iter()
-                                    .any(|needle| needle.find(taxid).is_some())
-                                {
+                                if needles.find(taxid).is_some() {
                                     batch_buffer.push(line);
                                     if batch_buffer.len()
                                         == batch_buffer.capacity()
@@ -126,6 +80,8 @@ where
                                                 "Send to writer failed: {e}"
                                             )
                                         })?;
+                                        batch_buffer =
+                                            Vec::with_capacity(batchsize)
                                     }
                                 }
                             }
@@ -201,108 +157,4 @@ where
         let _ = writer_handle.join().map_err(|_| "writer thread panicked")?;
         Ok(())
     })
-}
-
-fn read_sequence_id_from_koutput<P>(
-    file: P,
-    buffersize: usize,
-) -> std::result::Result<HashSet<Vec<u8>>, String>
-where
-    P: AsRef<Path> + Display,
-{
-    // Collect IDs into a HashSet for fast lookup
-    rprintln!("Extracting sequence IDs");
-    let opened =
-        File::open(file).map_err(|e| format!("Open file failed: {}", e))?;
-    let buffer = BufReader::with_capacity(buffersize, opened);
-    let id_set = buffer
-        .lines()
-        .filter_map(|line| {
-            line.ok().and_then(|str| {
-                // we selected the second column
-                str.split("\t").nth(1).and_then(|second| {
-                    // we remove empty sequence IDs
-                    if second.is_empty() {
-                        None
-                    } else {
-                        Some(second.as_bytes().to_vec())
-                    }
-                })
-            })
-        })
-        .collect::<HashSet<Vec<u8>>>();
-    Ok(id_set)
-}
-
-fn write_matching_reads<P>(
-    fq1: P,
-    ofile1: P,
-    fq2: Option<P>,
-    ofile2: Option<P>,
-    id_set: &HashSet<Vec<u8>>,
-    buffersize: usize,
-) -> std::result::Result<(), String>
-where
-    P: AsRef<Path> + Display,
-{
-    // Iterate all FASTQ records
-    rprintln!("Extracting the matching sequence from: {}", fq1);
-    write_matching_records(fq1, ofile1, buffersize, id_set)?;
-
-    if let (Some(in_file), Some(out_file)) = (fq2, ofile2) {
-        rprintln!("Extracting the matching sequence from: {}", in_file);
-        write_matching_records(in_file, out_file, buffersize, id_set)?;
-    }
-    Ok(())
-}
-
-fn write_matching_records<P>(
-    fastq: P,
-    fasta: P,
-    buffersize: usize,
-    id_set: &HashSet<Vec<u8>>,
-) -> std::result::Result<(), String>
-where
-    P: AsRef<Path> + Display,
-{
-    // Open input FASTQ
-    let in_file =
-        File::open(fastq).map_err(|e| format!("Open file failed: {}", e))?;
-    let in_buf = BufReader::with_capacity(buffersize, in_file);
-    let mut reader = Reader::new(in_buf);
-
-    // Open output FASTA
-    let out_file =
-        File::create(fasta).map_err(|e| format!("Open file failed: {}", e))?;
-    let out_buf = BufWriter::with_capacity(buffersize, out_file);
-    let mut writer = Writer::new(out_buf);
-
-    // Iterate all FASTQ records
-    for read in reader.records() {
-        let record =
-            read.map_err(|e| format!("Parse fastq record failed: {}", e))?;
-        if id_set.contains(<&[u8]>::from(record.name())) {
-            // convert FASTQ to FASTA record and write
-            let definition = definition::Definition::new(
-                record.name(),
-                Some(record.description().to_owned()),
-            );
-            let sequence =
-                sequence::Sequence::from(record.sequence().to_owned());
-            let fasta_record = Record::new(definition, sequence);
-            writer
-                .write_record(&fasta_record)
-                .map_err(|e| format!("Write fastq record failed: {}", e))?;
-        }
-    }
-    writer
-        .get_mut()
-        .flush()
-        .map_err(|e| format!("Flush lines failed: {}", e))?;
-    Ok(())
-}
-
-extendr_module! {
-    mod kractor;
-    fn kractor;
 }
